// sketched-app/src/app/enums/chord-type.enum.ts
export enum ChordType {
  FIFTH = 'FIFTH',
  TRIAD = 'TRIAD',
  SEVENTH = 'SEVENTH',
  NINTH = 'NINTH',
  ELEVENTH = 'ELEVENTH',
  THIRTEENTH = 'THIRTEENTH',
  ADDED_NINTH = 'ADDED_NINTH',
  SUS = 'SUS',
  SUS_SEVENTH = 'SUS_SEVENTH',
  SIXTH = 'SIXTH',
  SIXTH_NINTH = 'SIXTH_NINTH',
  ADDED_ELEVENTH = 'ADDED_ELEVENTH',
}